test: 생명주기 테스트 수정
