+++ di/src/test/java/com/example/fake/FakeActivity.kt
@@ -1,13 +1,13 @@
 package com.example.fake
 
-import FakeModule
+import FakeApplicationModule
 import FakeViewModel
 import com.example.di.DIActivity
 import com.example.di.DIModule
 import com.example.di.viewmodel.provideViewModel
 
 class FakeActivity : DIActivity() {
-    val viewModel: FakeViewModel by provideViewModel(FakeModule::class)
+    val viewModel: FakeViewModel by provideViewModel(FakeApplicationModule::class)
 
-    override val module: DIModule = FakeModule()
+    override val module: DIModule = FakeActivityModule()
 }
