+++ di/src/test/java/com/example/fake/FakeApplication.kt
@@ -1,9 +1,9 @@
 package com.example.fake
 
-import FakeModule
+import FakeApplicationModule
 import com.example.di.DIApplication
 import com.example.di.DIModule
 
 class FakeApplication : DIApplication() {
-    override val module: DIModule = FakeModule()
+    override val module: DIModule = FakeApplicationModule()
 }
