+++ di/src/test/java/com/example/fake/FakeViewModel.kt
@@ -1,6 +1,7 @@
 import com.example.di.annotation.Inject
 import com.example.di.annotation.Qualifier
 import com.example.di.annotation.QualifierType
+import com.example.fake.FakeActivityModule
 import com.example.fake.FakeCartRepository
 import com.example.fake.FakeDIViewModel
 import com.example.fake.FakeFieldRepository
@@ -9,7 +10,7 @@ import com.example.fake.FakeProductRepository
 class FakeViewModel(
     @Qualifier(QualifierType.Database) @Inject val fakeCartRepository: FakeCartRepository,
     @Inject val fakeProductRepository: FakeProductRepository,
-) : FakeDIViewModel(FakeModule::class) {
+) : FakeDIViewModel(FakeActivityModule::class) {
     @Inject
     lateinit var fakeFieldRepository: FakeFieldRepository
 }
